+++ src/lib.rs
@@ -18,6 +18,7 @@ pub mod modlist;
 pub mod modsource;
 pub mod modupdate;
 pub mod nexus;
+pub mod nxm;
 pub mod profile;
 pub mod publish;
 pub mod signing;